[BACKLOG-23288] Move consumer only meta fields from Jms Delegate to JmsConsumerMeta so they do not show up in jms producer mdi (#6055)
