+++ plugins/streaming/impls/jms/src/test/java/org/pentaho/di/trans/step/jms/JmsProducerTest.java
@@ -107,7 +107,7 @@ public class JmsProducerTest {
     jmsDelegate.amqUsername = "";
     jmsDelegate.amqPassword = "";
     jmsDelegate.destinationName = "testDestination";
-    jmsDelegate.receiveTimeout = "0";
+
 
     meta = new JmsProducerMeta( jmsDelegate );
